ocitrace.h: fix format errors in PerlIO_printf() calls
This code was trying to print two arguments that are likely to be size_t (as
the result of strlen() calls) using the "%d" format string. On systems where
size_t is wider than int (including most 64-bit systems), this will cause
undefined behaviour (which might in turn look like stack-smashing).