Move freeing global envhp to destruction of last dbh
As the order of object destruction during "global destruction" is
undefined. DBI traces show that the global environment is freed in the
driver destruction process before all database connections are closed
and their resources freed. This causes the segfault in #65.

This change moved freeing of the global environment handle from the
destruction of the driver handle (dbd_dr_destroy()) to the destruction
of the database handle (dbd_db_destroy()).

This implementation closes the global environment handle during the DBI
shutdown process if the last database handle uses the global
environment handle.

If the last database handles uses a session environment handle, the
global environment handle will not be destroyed.